Time: 1 ms (78.6%), Space: 10.7 MB (91.99%) - LeetHub

+++ 0071-simplify-path/0071-simplify-path.cpp
@@ -4,9 +4,9 @@ public:
         string x="",y="";
         vector<string>a;
         string ans="";
-        for(int i=0;i<(int)p.size();){
+        for(int i=0;i<p.size();){
             y="";
-            while(i<(int)p.size() && p[i]!='/'){
+            while(i<p.size() && p[i]!='/'){
                 y+=p[i];
                 i++;
             }
@@ -25,7 +25,7 @@ public:
                     y="";
                 }
             }
-            if(i<(int)p.size() && p[i]=='/') i++;
+            if(i<p.size() && p[i]=='/') i++;
         }
         if(!x.empty()) a.push_back(x);
         for(int i=0;i<a.size();i++){
